\name{prediction_res}
\alias{prediction_res}
\docType{data}
\title{
Prediction results for V1 voxels as generated by the Gallant lab in UC Berkeley and published in Kay et al. (2008). 
}
\description{
The correlation between measured response and predicted response on validation data for 1250 V1 voxels. 
The prediction algorithms are described in detail in Kay et al, 2008. We compare these prediction results 
to the explainable variance estimated with the shuffle estimator. 
}
\usage{data(prediction_res)}
\format{
A numerical vector of correlation coefficients.
}
\source{
Kay, Naselaris, Prenger and Gallant (2008), "Identifying natural images from human brain activity"
}
\references{
Benjamini and Yu (2013)
}
\keyword{datasets}
